feat: add unit roster and loadout config screens with class skill rework
- Replace 7_Unit/8_ConfigUnit scenes with 7_UnitList/8_UnitConfig, driven by new
  UnitListManager, UnitConfigManager, UnitListStateMachine, and OptionListStateMachine
- Add DTOs.cs, WeaponDefinitionRegistry, and TrinketDefinitionRegistry for backend-driven
  unit/weapon/trinket data, plus UnitBar, UnitContainer, SlotConfig, and OptionItem UI components
- Replace SkillGuard class skill with per-class skills (Assassin, Tank, and placeholders)
  and add matching ClassSkillAssasinEffectClient/Server effect pair
- Add LoadingScreen and EyesCanvasUI, and import supporting UI art (baseFrame, baseMask,
  DropDownScroll, EYEs, HomeIcon, ShoeIcon)
- Rework client/server match sync (ClientMatchSession, ClientSyncMachine,
  ClientVisualController, ServerMatchSession, ServerTimelineManager) to support the new
  loadout flow

+++ Assets/Scripts/Server/ServerBackendClient.cs
@@ -53,8 +53,8 @@ public class UnitGradeStatsData
     public int maxHP;
     public int maxSkillPoint;
     public int speed;
-    public float damageMultiplier;
-    public float damageReduction;
+    public int damageMultiplier;
+    public int damageReduction;
 }
 
 [System.Serializable]
@@ -65,8 +65,8 @@ public class EquippedEquipmentData
     public int maxHP;
     public int maxSkillPoint;
     public int speed;
-    public float damageMultiplier;
-    public float damageReduction;
+    public int damageMultiplier;
+    public int damageReduction;
 }
 
 public class ServerBackendClient : MonoBehaviour
